requirements.txt
Django==1.6.5
gevent==1.0.1
gevent-socketio==0.3.6
gevent-websocket==0.9.3
greenlet==0.4.2
gunicorn==19.1.0
wsgiref==0.1.2
